requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via pytest
coverage==4.5.1           # via pytest-cov
django-model-utils==3.1.2
more-itertools==4.3.0     # via pytest
pathlib2==2.3.2           # via pytest
pluggy==0.7.1             # via pytest
py==1.5.4                 # via pytest
pytest-cov==2.5.1
pytest-django==3.3.3
pytest==3.7.1             # via pytest-cov, pytest-django
pytz==2018.5              # via django
six==1.11.0               # via more-itertools, pathlib2, pytest
